update examples after moving component import to top level in cl/270088663
PiperOrigin-RevId: 276755744
